# Optimize compile map