… MapAnalyzer/Polygon.py
… MapAnalyzer/Region.py
… MapAnalyzer/__init__.py
… MapAnalyzer/cext/wrapper.py
… MapAnalyzer/constants.py
… MapAnalyzer/constructs.py
… MapAnalyzer/decorators.py
… MapAnalyzer/destructibles.py
… MapAnalyzer/exceptions.py
… MapAnalyzer/utils.py
… dummybot.py
… tests/test_suite.py